fix sync
